// pronti-frontend-main/front/src/app/modulos/abono-prestamo/form/form-abono-prestamo.enum.ts
export enum FormAbonoPrestamoEnum {
  fechaPago = 'fechaPago',
  numeroPago = 'numeroPago',
  modalidadDescuento = 'modalidadDescuento',
  mes = 'mes',
  anio = 'anio',
  valorCuota = 'valorCuota',
  estaPagado = 'estaPagado',
  valorCapital = 'valorCapital',
  valorTasaInteres = 'valorTasaInteres',
}